+++ rep_score_portal/config.py
@@ -42,11 +42,11 @@ DEI_CREATIVE_REVIEWS_LABEL_2 = (
     'Could the work be reinforcing negative stereotypes?'
 )
 DEI_CREATIVE_REVIEWS_LABEL_3 = (
-    'What cultural references could we be misappropriating?'
+    'What cultural references could we be appropriating or misrepresenting?'
 )
 DEI_CREATIVE_REVIEWS_LABEL_4 = (
-    'Can we be more inclusive with regard to age, body type, disability, ethnicity, gender, sexual '
-    'orientation?'
+    'Can we be more inclusive with regard to age (older adults), body type (fat characters), '
+    'disability, race/ethnicity, skin tone, gender, sexuality?'
 )
 DEI_CREATIVE_REVIEWS_LABEL_5 = (
     'How can we, and our clients, make choices that lead to more inclusive and equitable work?'